client/requirements.txt
hazelcast-python-client==3.12.2